+++ scripts/genl10n.js
@@ -10,22 +10,22 @@ const messages = read.sync(process.argv[2]);
 const targetDir = process.argv[3];
 const langs = ['de', 'en', 'es', 'fr', 'gl', 'hu', 'it', 'nl','pl'];
 
-data = [];
-data.push("<?php");
-data.push("# Generated file. Please don\'t edit here,");
-data.push("# edit messages.yml instead. ");
-data.push("#");
-
 langs.forEach(lang => {
-    data.push(`$lang${lang} = array(`);
+    data = [];
+    data.push("<?php\n");
+    data.push("# Generated file. Please don\'t edit here,");
+    data.push("# edit messages.yml instead.");
+    data.push("#");
+    data.push(`$messages = [`);
     Object.entries(messages).forEach( msgArr => {
         const msg = msgArr[0];
         const locs = msgArr[1];
         if (locs[lang] !== undefined) {
-            data.push(`\'${msg}\' => \'${locs[lang]}\',`);
+            data.push(`    \'${msg}\' => \'${locs[lang]}\',`);
         }
     });
-    data.push(");\n");
+    data.push("];\n");
+    data.push(`return $messages;`);
+    data.push("");
+    fs.writeFileSync(`${targetDir}/messages.${lang}.php`, data.join("\n"));
 });
-
-fs.writeFileSync(`${targetDir}/langs.php`, data.join("\n"));