+++ tests/MiddlewareTest.php
@@ -62,8 +62,6 @@ class MiddlewareTest extends PHPUnit\Framework\TestCase
 
     public function getApp($login = 0)
     {
-        global $langen;
-        require('config/langs.php');
         $app = require(dirname(__DIR__) . '/config/bootstrap.php');
         $settings = $app->getContainer()->get('globalSettings');
         $settings->must_login = $login;